app/requirements.txt
flask==2.2.3
flask-wtf==1.1.1
torch==2.0.0
librosa==0.10.0.post2
soundfile==0.12.1
google-cloud-storage==2.6.0
mlflow==2.2.1
numpy==1.24.3